perf: make index join/merge handlers optional
Restore the classes `JoinHandler` and `PropertiesMergeHandler` as the default
handlers for the transformation functions Join, Groovy Join, Merge and Groovy
Merge.

The previous default handlers `IndexJoinHandler` and `IndexMergeHandler` have
demonstrated negative impact on the performance of some transformations. Until
this is better understood and fixed, these handlers are now optional and can be
activated via the following Java properties and environment variables:

- Join/Groovy Join: `hale.functions.use_index_join_handler` / `HALE_FUNCTIONS_USE_INDEX_JOIN_HANDLER`
- Merge/Groovy Merge: `hale.functions.use_index_merge_handler` / `HALE_FUNCTIONS_USE_INDEX_MERGE_HANDLER`

ING-4464