fix: download all Shapefiles when given URL contains FIX file
If source location is provided as URL and it contains *.fix file,
then Geotools throws an exception reading the files.
One solution to this is to download all the files to a temporary location
incase *.fix file was provided in the URL.
This would prevent Geotools throwing Read Channel IOException.

ING-4811